Messed With Stats Per Level And Added A 4th, Unique Action For Each Job
Played around with the stats the player games per level and their starting stats, not finished.
Mostly worked on implementing a 4th action for each job and creating the groundwork for any future additions specifically for jobs. We can move dodge and run around if we want, and/or, we can change it to what I initially named the actions for each class if we so desire.
Changed neededExperience per level to player->level * 50, but we can make it go up as they reach a certain level, say 4.
Increased damage range as the player levels up to 4 and beyond.